+++ Sources/JWTKeychain/Models/JWTKeychainMiddlewares.swift
@@ -1,0 +1,6 @@
+import Vapor
+
+public struct JWTKeychainMiddlewares<U: JWTKeychainUserType>: Service {
+    public let accessMiddlewares: [Middleware]
+    public let refreshMiddlewares: [Middleware]?
+}
