+++ Sources/JWTKeychain/Models/JWTKeychainUser/JWTKeychainUser+JWTKeychainUserType.swift
