+++ Sources/JWTKeychain/Models/JWTKeychainUser/JWTKeychainUserType.swift
