+++ Sources/JWTKeychain/Provider/JWTKeychainProvider.swift
@@ -1,14 +1,6 @@
 import Authentication
-import Fluent
-import JWT
 import Service
 import Sugar
-import Vapor
-
-public struct JWTKeychainMiddlewares<U: JWTKeychainUserType>: Service {
-    public let accessMiddlewares: [Middleware]
-    public let refreshMiddlewares: [Middleware]?
-}
 
 public final class JWTKeychainProvider<U: JWTKeychainUserType> {
     public let middlewares: JWTKeychainMiddlewares<U>
@@ -50,41 +42,5 @@ extension JWTKeychainProvider: Provider {
 
     public func didBoot(_ container: Container) throws -> Future<Void> {
         return .done(on: container)
-    }
-}
-
-// MARK: Routes
-
-public extension Router {
-    func useJWTKeychainRoutes<U: JWTKeychainUserType>(
-        _ type: U.Type,
-        on container: Container
-    ) throws {
-        let config: JWTKeychainConfig<U> = try container.make()
-        let middlewares: JWTKeychainMiddlewares<U> = try container.make()
-        let access = self.grouped(middlewares.accessMiddlewares)
-
-        if let registerPath = config.endpoints.register {
-            self.post(registerPath, use: config.controller.register)
-        }
-
-        if let loginPath = config.endpoints.login {
-            self.post(loginPath, use: config.controller.logIn)
-        }
-
-        if let mePath = config.endpoints.me {
-            access.get(mePath, use: config.controller.me)
-        }
-
-        if let updatePath = config.endpoints.update {
-            access.patch(updatePath, use: config.controller.update)
-        }
-
-        if
-            let refreshMiddlewares = middlewares.refreshMiddlewares,
-            let tokenPath = config.endpoints.token
-        {
-            self.grouped(refreshMiddlewares).post(tokenPath, use: config.controller.token)
-        }
     }
 }
